Deduplicate password reset email flow
